agents-portal/agents/contexto-midias/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.58.1
python-dotenv==1.2.0
pdfminer.six==20231228
docling==2.58.0
filetype==1.2.0
requests>=2.32.2
